refactor: template_vocabulary 도메인 분리 #151
