+++ api/src/main/java/vook/server/api/devhelper/app/InitService.java
@@ -6,14 +6,14 @@ import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.demo.model.DemoTerm;
 import vook.server.api.domain.demo.model.DemoTermRepository;
 import vook.server.api.domain.demo.model.DemoTermSynonymRepository;
+import vook.server.api.domain.template_vocabulary.logic.TemplateVocabularyLogic;
+import vook.server.api.domain.template_vocabulary.logic.dto.TemplateVocabularyCreateCommand;
+import vook.server.api.domain.template_vocabulary.model.TemplateTermRepository;
+import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyRepository;
+import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.user.model.SocialUserRepository;
 import vook.server.api.domain.user.model.UserInfoRepository;
 import vook.server.api.domain.user.model.UserRepository;
-import vook.server.api.domain.vocabulary.logic.TemplateVocabularyLogic;
-import vook.server.api.domain.vocabulary.logic.dto.TemplateVocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.TemplateTermRepository;
-import vook.server.api.domain.vocabulary.model.TemplateVocabularyRepository;
-import vook.server.api.domain.vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.vocabulary.model.TermRepository;
 import vook.server.api.infra.search.demo.MeilisearchDemoTermSearchService;
 import vook.server.api.infra.vocabulary.cache.UserVocabularyCacheRepository;
