+++ api/src/main/java/vook/server/api/domain/template_vocabulary/logic/TemplateVocabularyLogic.java
@@ -1,9 +1,9 @@
-package vook.server.api.domain.vocabulary.logic;
+package vook.server.api.domain.template_vocabulary.logic;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import vook.server.api.domain.vocabulary.logic.dto.TemplateVocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.*;
+import vook.server.api.domain.template_vocabulary.logic.dto.TemplateVocabularyCreateCommand;
+import vook.server.api.domain.template_vocabulary.model.*;
 import vook.server.api.globalcommon.annotation.DomainLogic;
 
 import java.util.List;
