+++ api/src/test/java/vook/server/api/domain/template_vocabulary/logic/TemplateVocabularyLogicTest.java
@@ -1,11 +1,11 @@
-package vook.server.api.domain.vocabulary.logic;
+package vook.server.api.domain.template_vocabulary.logic;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.vocabulary.logic.dto.TemplateVocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.*;
+import vook.server.api.domain.template_vocabulary.logic.dto.TemplateVocabularyCreateCommand;
+import vook.server.api.domain.template_vocabulary.model.*;
 import vook.server.api.testhelper.IntegrationTestBase;
 
 import java.util.List;
