+++ api/src/test/java/vook/server/api/testhelper/creator/TestTemplateVocabularyCreator.java
@@ -3,9 +3,9 @@ package vook.server.api.testhelper.creator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.vocabulary.logic.TemplateVocabularyLogic;
-import vook.server.api.domain.vocabulary.logic.dto.TemplateVocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.TemplateVocabularyType;
+import vook.server.api.domain.template_vocabulary.logic.TemplateVocabularyLogic;
+import vook.server.api.domain.template_vocabulary.logic.dto.TemplateVocabularyCreateCommand;
+import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
 
 import java.util.List;
