+++ api/src/test/java/vook/server/api/web/user/usecase/OnboardingUserUseCaseTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
+import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.user.model.Funnel;
 import vook.server.api.domain.user.model.Job;
 import vook.server.api.domain.user.model.UserRepository;
-import vook.server.api.domain.vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.vocabulary.model.TermRepository;
 import vook.server.api.domain.vocabulary.model.UserUid;
 import vook.server.api.domain.vocabulary.model.VocabularyRepository;
